requirements.txt
fastapi
uvicorn
pydantic
langchain
faiss-cpu
jinja2
numpy
python-dotenv
requests
nomic 
torch 
transformers